FilterState documentation
