+++ Sources/FilterState/Accessors/HierarchicalGroupAccessor.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 
+/// Provides a convenient interface to a hierarchical group contained by FilterState
 public struct HierarchicalGroupAccessor: SpecializedGroupAccessor {
   
   public typealias Filter = InstantSearchCore.Filter.Facet
