+++ Sources/FilterState/FilterState+DisjunctiveFaceting.swift
@@ -1,0 +1,17 @@
+//
+//  FilterState+DisjunctiveFaceting.swift
+//  InstantSearchCore
+//
+//  Created by Vladislav Fitc on 05/08/2019.
+//  Copyright © 2019 Algolia. All rights reserved.
+//
+
+import Foundation
+
+extension FilterState: DisjunctiveFacetingDelegate {
+  
+  public var disjunctiveFacetsAttributes: Set<Attribute> {
+    return filters.disjunctiveFacetsAttributes
+  }
+  
+}
